feat: endpoint handler
